+++ compiler/astdef.nim
@@ -8,7 +8,7 @@
 #
 
 import
-  lineinfos, options, ropes, idents, int128
+  lineinfos, options, ropes, idents, int128, wordrecg
 
 import std/[tables, hashes]
 
@@ -125,6 +125,8 @@ type
     sfCodegenDecl     # type, proc, global or proc param is marked as codegenDecl
     sfWasGenSym       # symbol was 'gensym'ed
     sfForceLift       # variable has to be lifted into closure environment
+
+    sfHasDeferredPragmas  # field has one or more deferred pragma expressions
 
     sfDirty           # template is not hygienic (old styled template) module,
                       # compiled from a dirty-buffer
@@ -397,6 +399,7 @@ type
     tfIsOutParam
     tfSendable
     tfImplicitStatic
+    tfHasDeferredPragmas  # type has one or more deferred pragma expressions
 
   TTypeFlags* = set[TTypeFlag]
 
@@ -589,6 +592,13 @@ type
   TNodeSeq* = seq[PNode]
   PType* = ref TType
   PSym* = ref TSym
+
+  DeferredPragmaExpr* = object
+    ## A pragma expression that needs evaluation during generic instantiation.
+    ## Used for pragmas like size that can reference generic params.
+    word*: TSpecialWord   ## which pragma (wSize, etc.)
+    expr*: PNode          ## the deferred expression containing generic params
+
   TNode*{.final, acyclic.} = object # on a 32bit machine, this takes 32 bytes
     when defined(useNodeIds):
       id*: int
@@ -709,6 +719,7 @@ type
       guardImpl*: PSym
       bitsizeImpl*: int
       alignmentImpl*: int # for alignment
+      deferredExprsImpl*: seq[DeferredPragmaExpr]  # deferred pragma expressions (nil = none)
     else: nil
     magicImpl*: TMagic
     typImpl*: PType
@@ -796,6 +807,7 @@ type
                               # -1 means that the size is unknown
     alignImpl*: int16             # the type's alignment requirements
     paddingAtEndImpl*: int16      #
+    deferredExprsImpl*: seq[DeferredPragmaExpr]  # deferred pragma expressions (nil = none)
     locImpl*: TLoc
     typeInstImpl*: PType          # for generic instantiations the tyGenericInst that led to this
                               # type.